requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aioredis==1.3.1
    # via channels-redis
amqp==5.0.5
    # via kombu
asgiref==3.3.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
async-timeout==3.0.1
    # via aioredis
attrs==20.3.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==21.3.1
    # via daphne
automat==20.2.0
    # via twisted
billiard==3.6.3.0
    # via celery
celery==5.0.5
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-celery-results
cffi==1.14.5
    # via cryptography
channels-redis==3.2.0
    # via -r requirements.in
channels==3.0.3
    # via
    #   -r requirements.in
    #   channels-redis
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
constantly==15.1.0
    # via twisted
cryptography==3.4.6
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==3.0.1
    # via channels
django-celery-beat==2.2.0
    # via -r requirements.in
django-celery-results==2.0.1
    # via -r requirements.in
django-timezone-field==4.1.1
    # via django-celery-beat
django==3.1.7
    # via
    #   -r requirements.in
    #   channels
    #   django-celery-beat
    #   django-timezone-field
    #   djangorestframework
djangorestframework==3.12.2
    # via -r requirements.in
gunicorn==20.0.4
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
hiredis==1.1.0
    # via aioredis
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.1
    # via
    #   hyperlink
    #   twisted
importlib-metadata==3.7.2
    # via kombu
incremental==21.3.0
    # via twisted
kombu==5.0.2
    # via celery
msgpack==1.0.2
    # via channels-redis
prompt-toolkit==3.0.17
    # via click-repl
psycopg2==2.8.6
    # via -r requirements.in
pyasn1-modules==0.2.8
    # via service-identity
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pycparser==2.20
    # via cffi
pyopenssl==20.0.1
    # via twisted
python-crontab==2.5.1
    # via django-celery-beat
python-dateutil==2.8.1
    # via python-crontab
pytz==2021.1
    # via
    #   celery
    #   django
    #   django-timezone-field
pyyaml==5.4.1
    # via -r requirements.in
redis==3.5.3
    # via -r requirements.in
service-identity==18.1.0
    # via twisted
six==1.15.0
    # via
    #   automat
    #   click-repl
    #   pyopenssl
    #   python-dateutil
sqlalchemy==1.3.23
    # via -r requirements.in
sqlparse==0.4.1
    # via django
twisted[tls]==21.2.0
    # via daphne
txaio==21.2.1
    # via autobahn
typing-extensions==3.7.4.3
    # via
    #   importlib-metadata
    #   uvicorn
uvicorn==0.13.4
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
wcwidth==0.2.5
    # via prompt-toolkit
zipp==3.4.1
    # via importlib-metadata
zope.interface==5.2.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
